requirements.txt
notebooklm-py>=0.3.4
httpx>=0.27.0
playwright>=1.44.0
